Ajout d'un composant pour mutualiser la logique et ajout des icones dans le menu gauche
