+++ front/gatsby/src/components/ReferenceTypeIcon.scss
@@ -1,0 +1,3 @@
+.icon {
+    vertical-align: middle;
+}
