+++ front/gatsby/src/components/Write/Reference.js
@@ -2,9 +2,10 @@ import React,{useState, useEffect} from 'react'
 import {CopyToClipboard} from 'react-copy-to-clipboard';
 
 import styles from './reference.module.scss'
+import ReferenceTypeIcon from '../ReferenceTypeIcon'
 
 export default ({ entry }) => {
-  const {key, title} = entry
+  const {key, title, type} = entry
 
   const [copied, setCopied] = useState(false)
   useEffect(()=>{
@@ -16,6 +17,7 @@ export default ({ entry }) => {
   return (
     <CopyToClipboard text={`[@${key}]`} onCopy={() => setCopied(true)}>
       <p className={`${styles.reference} ${copied ? styles.clicked: null}`} title={title}>
+        <ReferenceTypeIcon type={type}/>
         <span>@{key}</span>
       </p>
     </CopyToClipboard>
