+++ front/gatsby/src/helpers/bibtex.js
@@ -20,12 +20,11 @@ export function toBibtex (entries) {
 }
 
 /**
- *
  * We tolerate `unexpected_field` warnings as it's user provided, it does not have any side effect
  * @see https://github.com/EcrituresNumeriques/stylo/issues/187
  *
  * @param {string} bibtext
- * @returns {Promise[{success: number,empty: boolean,warnings: Array.<string>,error: Array.<string>}]}
+ * @returns {Promise<{success: number,empty: boolean,warnings: Array.<string>,error: Array.<string>}>}
  */
 export function validate(bibtext) {
   return parse(bibtext).then(result => ({
@@ -36,4 +35,33 @@ export function validate(bibtext) {
         .filter(error => error.type !== 'unexpected_field')
         .map(error => error.type + ' at line ' + error.line)
   }))
+}
+
+const IconNameMap = {
+  article: 'journal-article',
+  book: 'book',
+  booklet: 'journal-article',
+  inbook: 'book-section',
+  incollection: 'document',
+  inproceedings: 'conference-paper',
+  manual: 'book',
+  mastersthesis: 'thesis',
+  misc: 'journal-article',
+  phdthesis: 'thesis',
+  proceedings: 'book',
+  techreport: 'report',
+  unpublished: 'manuscript'
+}
+
+/**
+ * Get the icon name for a given Bibtex type.
+ * @param bibtextType
+ * @returns {string}
+ */
+export function iconName (bibtextType) {
+  const iconName = IconNameMap[bibtextType]
+  if (iconName) {
+    return iconName
+  }
+  return 'book'
 }
